src/requirements.txt
Django
unicorn
uvicorn
python-dotenv
psycopg2-binary